Implemented headless testing
