--- tests/test_algorithm/test_configuration/test_graphics/compare_ss.py
@@ -1,7 +1,0 @@
-import numpy as np
-
-# Lower values better
-def mse(imageA, imageB):
-    err = np.sum((imageA.astype("float") - imageB.astype("float")) ** 2)
-    err /= float(imageA.shape[0] * imageA.shape[1])
-    return err
